moved components into seperate files. created test file and folder
