+++ src/__tests__/game.test.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import { Game } from '../game';
+
+ it('renders without crashing', () => {
+  shallow(<Game />);
+});
